Extract auto-update pipeline into AppUpdater library (#1824, #1825, #1826, #1827)
Replace the RunBotCore-embedded AutoUpdater with a standalone, host-agnostic
AppUpdater package target carrying zero RunBotCore dependency.

- P2 (#1824): AutoUpdater caseless enum → @MainActor final class AppUpdater with
  injected repo/currentVersion/assetName/schedulerIdentifier/userDefaults/
  betaChannelProvider. isInstalling/isDownloading become instance flags; the
  NSBackgroundActivityScheduler is retained per-instance and invalidated in
  deinit. All RunBot strings ("runbot-hq/run-bot", "RunBot.zip",
  "io.github.runbot-hq.update-check") move to the AppDelegate call site.
- P3 (#1825): new @MainActor UpdateStateProviding protocol (read-only props +
  named mutation methods); RunnerState conforms via RunnerState+AppUpdater.swift
  and gains setDownloadStarted/setDownloadComplete/setUpdateFailed; rehydrate
  now clears updateActionFailed.
- P4 (#1826): move+rename AutoUpdater+*.swift → Sources/AppUpdater/AppUpdater+*;
  betaChannel read via a MainActor hop; module-local os.Logger in
  AppUpdater+Logging.swift; NSBackgroundActivityScheduler + AppKit usage guarded
  by #if canImport(AppKit); runCommand refactored to a @concurrent free function
  (no withCheckedContinuation/DispatchQueue). Bundle+Version stays in RunBotCore
  and imports AppUpdater for isNewer.
- P5 (#1827): AppUpdaterDefaults scoped via init(domain:) (domain =
  schedulerIdentifier); cache dir/filename de-RunBot-ified; AppDelegate startup
  migrated to rehydrateCachedUpdateIfNewer / checkAndHandle / scheduleBackgroundCheck.

De-RunBot-ified the install path to locate the first *.app in the archive
(instead of a hardcoded RunBot.app) and derive the backup name from the running
bundle.
