--- Sources/AppUpdater/.gitkeep
@@ -1,1 +1,0 @@
-# AppUpdater target — see #1821 for implementation plan
