+++ Sources/AppUpdater/AppUpdater+ProcessHelper.swift
@@ -1,0 +1,84 @@
+// AppUpdater+ProcessHelper.swift
+// AppUpdater
+import Foundation
+
+// MARK: - Process helper
+
+/// Runs a command and returns `true` on exit code 0, `false` otherwise.
+///
+/// Used for `ditto` (unzip and bundle replacement) which is short-lived and
+/// does not need streaming output.
+///
+/// ## `@concurrent` — the blocking wait runs off every actor executor
+///
+/// This is a `@concurrent` async free function, so its synchronous body
+/// (`process.run()` + `waitUntilExit()`) runs on the cooperative thread pool's
+/// concurrent executor, never on an actor serial executor (Pillar 5,
+/// `docs/architecture/concurrency-overview.md`). The previous implementation
+/// wrapped the work in `withCheckedContinuation` + `DispatchQueue.global` to
+/// hop off the main thread; `@concurrent` expresses the same intent directly and
+/// removes the manual GCD plumbing.
+///
+/// The tradeoff: `waitUntilExit()` holds one cooperative-pool thread for the
+/// lifetime of the subprocess (~1–2 s for `ditto`). For a two-call sequence that
+/// runs once per update cycle this is completely acceptable; there is no
+/// scenario where many concurrent `ditto` invocations would exhaust the pool
+/// (`isInstalling` serialises the install path).
+///
+/// ## Why stderr is captured (not discarded)
+///
+/// Routing stderr to `nullDevice` made `ditto` failures completely silent: the
+/// install path would fail, `updateActionFailed` would flip to `true`, and the
+/// user would see a "Download" fallback with no indication of what went wrong.
+/// Stderr is now piped and logged at error level on failure so that Console.app
+/// and crash reports contain actionable information. Stdout remains discarded —
+/// `ditto` produces no useful stdout.
+///
+/// ## ⚠️ `waitUntilExit()` BEFORE `readDataToEndOfFile()` — NOT A PIPE DEADLOCK ⚠️
+///
+/// Automated reviewers sometimes flag this ordering as a potential pipe-buffer
+/// deadlock. It is NOT a deadlock risk here. The classic deadlock requires the
+/// subprocess to write ≥ ~64 KB to stderr *before* exiting — filling the OS pipe
+/// buffer and blocking mid-write. That cannot arise for `ditto`: on success it
+/// writes nothing to stderr; on failure it emits a single short error line
+/// (well under 1 KB). The buffer never fills; `ditto` exits immediately; the
+/// pipe drains normally.
+///
+/// REVIEWER: Do NOT refactor this to a concurrent drain on the basis of a
+/// generic pipe-deadlock warning. The warning does not apply to `ditto`. If this
+/// helper is ever extended to run a different executable that CAN produce large
+/// stderr output, add the concurrent drain at that point and update this comment.
+@concurrent
+func runCommand(_ executable: String, args: [String]) async -> Bool {
+    let process = Process()
+    process.executableURL = URL(fileURLWithPath: executable)
+    process.arguments = args
+    process.standardOutput = FileHandle.nullDevice
+
+    // Pipe stderr so failures are diagnosable. stdout is still discarded —
+    // ditto produces no useful stdout on success.
+    let stderrPipe = Pipe()
+    process.standardError = stderrPipe
+
+    do {
+        try process.run()
+        // waitUntilExit() before readDataToEndOfFile() is safe here — ditto's
+        // stderr output is always < 1 KB, so the OS pipe buffer (64 KB) is never
+        // filled and no deadlock can occur. See the doc comment above.
+        process.waitUntilExit()
+        let succeeded = process.terminationStatus == 0
+        if !succeeded {
+            let stderrData = stderrPipe.fileHandleForReading.readDataToEndOfFile()
+            let stderrMsg = String(data: stderrData, encoding: .utf8) ?? "(unreadable)"
+            appUpdaterLogger.error(
+                "\(executable, privacy: .public) failed (exit \(process.terminationStatus, privacy: .public)): \(stderrMsg, privacy: .public)"
+            )
+        }
+        return succeeded
+    } catch {
+        appUpdaterLogger.error(
+            "could not launch \(executable, privacy: .public): \(error.localizedDescription, privacy: .public)"
+        )
+        return false
+    }
+}
