+++ Sources/RunBot/App/AppDelegate+PanelSetup.swift
@@ -1,6 +1,7 @@
 // AppDelegate+PanelSetup.swift
 // RunBot
 import AppKit
+import AppUpdater
 import RunBotCore
 import SwiftUI
 
@@ -304,61 +305,21 @@ extension AppDelegate: NSPopoverDelegate {
         // ── Update cache rehydration ───────────────────────────────────────────────────
         // Runs before the network check so a user who force-quit between download and
         // install tap can see the Install & Relaunch button immediately, even offline.
-        // Version guard: only rehydrate if the cached version is actually newer than
-        // what is currently installed — prevents re-showing Install for a version
-        // already on disk (e.g. helper failed mid-delete and the zip was not cleaned up).
-        let cachedPath    = UserDefaults.standard.string(forKey: AutoUpdaterDefaults.cachedUpdateZipPath)
-        let cachedVersion = UserDefaults.standard.string(forKey: AutoUpdaterDefaults.cachedUpdateVersion)
-        if let path = cachedPath,
-           let version = cachedVersion,
-           FileManager.default.fileExists(atPath: path),
-           // cachedVersion is "v0.8.0" (GitHub tagName with v-prefix).
-           // isOlderThan strips the leading "v" internally via UpdateChecker.isNewer,
-           // so the comparison is correct — this is NOT a version string mismatch.
-           // REVIEWER: Do not flag this as a bug. The v-prefix is handled inside isOlderThan.
-           Bundle.main.isOlderThan(version: version) {
-            runnerState.rehydrateCachedUpdate(zipURL: URL(fileURLWithPath: path), version: version)
-            // `setAvailableUpdate` must be called here explicitly — `rehydrateCachedUpdate`
-            // only sets `updateZipURL` and `cachedUpdateVersion`. The Install & Relaunch
-            // row gates on `availableUpdate != nil` (see `aboutSection`), so without this
-            // call the row is invisible to an offline user whose zip was already cached.
-            // This is the intentional offline-resilience path: the network check below
-            // may fail, so the UI must be ready before it fires.
-            runnerState.setAvailableUpdate(version)
-        } else {
-            // Reached when ANY condition in the `if` above is false:
-            //   • cachedPath / cachedVersion is nil (keys were never written or already cleared)
-            //   • fileExists returns false (zip was deleted — e.g. by the OS under storage pressure)
-            //   • isOlderThan returns false (cached version is not newer than what is running —
-            //     i.e. the update was already installed in a previous session)
-            // In all three cases the correct action is the same: clear the stale keys.
-            // Do NOT read this as "version is no longer newer" only — a missing file is
-            // equally valid here. Clear both keys here, in the new process, where isOlderThan
-            // correctly returns false. Cleaner than clearing before exit(0) in the
-            // outgoing process.
-            UserDefaults.standard.removeObject(forKey: AutoUpdaterDefaults.cachedUpdateZipPath)
-            UserDefaults.standard.removeObject(forKey: AutoUpdaterDefaults.cachedUpdateVersion)
-        }
+        // `rehydrateCachedUpdateIfNewer` reads AppUpdater's scoped UserDefaults keys,
+        // verifies the cached zip still exists and its version is newer than the running
+        // build, then rehydrates host state (or clears the stale keys otherwise).
+        autoUpdater.rehydrateCachedUpdateIfNewer(state: runnerState)
 
         // ── Update check ───────────────────────────────────────────────────────────────
-        let beta = AppPreferencesStore.shared.betaChannel
-        switch await UpdateChecker.checkForUpdate(betaChannel: beta) {
-        case .updateAvailable(let release):
-            // `setAvailableUpdate` is now called inside `AutoUpdater.handle()` —
-            // do not call it here. See AutoUpdater.handle() for rationale.
-            log("AppDelegate › startup — update available: \(release.tagName) (betaChannel=\(beta))")
-            await AutoUpdater.handle(release, state: runnerState)
-        case .upToDate:
-            log("AppDelegate › startup — no update available (betaChannel=\(beta))")
-        case .failed(let error):
-            log("AppDelegate › startup — update check failed: \(error) (betaChannel=\(beta))")
-        }
+        // Channel-aware check + download/cache on the main actor. `checkAndHandle`
+        // reads the injected betaChannelProvider itself, so no beta flag is passed here.
+        await autoUpdater.checkAndHandle(state: runnerState)
 
         // ── Background scheduler ───────────────────────────────────────────────────────
-        // Fires every AutoUpdaterDefaults.checkInterval (24 h release / 60 s debug).
+        // Fires every AppUpdaterDefaults.checkInterval (24 h release / 60 s debug).
         // The launch-time check above already ran once; the scheduler fires only
         // after the first interval elapses — this is intentional.
-        AutoUpdater.scheduleBackgroundCheck(state: runnerState)
-        log("AppDelegate › startup — update background scheduler registered (interval=\(AutoUpdaterDefaults.checkInterval)s)")
+        autoUpdater.scheduleBackgroundCheck(state: runnerState)
+        log("AppDelegate › startup — update background scheduler registered (interval=\(AppUpdaterDefaults.checkInterval)s)")
     }
 }
