+++ Sources/RunBotCore/Runner/Models/RunnerState.swift
@@ -18,8 +18,10 @@ import Observation
 /// `{ get set }` requirement — see `RunnerViewModelProtocol` for the rationale.
 /// Only `LocalRunnerStore` (in `RunBotCore`) writes them in practice.
 /// The auto-update download properties (`updateZipURL`, `cachedUpdateVersion`,
-/// `updateAssetMissing`, `updateActionFailed`) are `public internal(set)` — only
-/// `AutoUpdater` (same `RunBotCore` module) writes them via `await MainActor.run`.
+/// `updateAssetMissing`, `updateActionFailed`) are `public internal(set)` for
+/// reads, but are written by `AppUpdater` (in the `AppUpdater` module) through
+/// the `UpdateStateProviding` mutation methods below. `RunnerState` conforms to
+/// that protocol in `RunnerState+AppUpdater.swift`.
 /// Views and app-layer code are read-only consumers of all properties.
 @Observable
 @MainActor
@@ -89,7 +91,7 @@ public final class RunnerState {
     /// default argument fails to compile.
     public init() {}
 
-    // MARK: - Auto-update state (pushed by AutoUpdater)
+    // MARK: - Auto-update state (driven by AppUpdater via UpdateStateProviding)
 
     /// The latest available version string if a newer version exists, or `nil` if
     /// up to date.
@@ -99,9 +101,9 @@ public final class RunnerState {
 
     /// Sets `availableUpdate`.
     ///
-    /// Called from `AutoUpdater.handle(_:state:)` on every `.updateAvailable` result
+    /// Called from `AppUpdater.handle(_:state:)` on every `.updateAvailable` result
     /// (including the launch-time check in `AppDelegate+PanelSetup`) and from
-    /// `AutoUpdater.scheduleBackgroundCheck` to clear a stale row on `.upToDate`
+    /// `AppUpdater.scheduleBackgroundCheck` to clear a stale row on `.upToDate`
     /// or `.failed` results (when no zip is cached).
     ///
     /// Using an explicit method (rather than direct property assignment) keeps
@@ -122,11 +124,37 @@ public final class RunnerState {
 
     /// Rehydrates cached download state from `UserDefaults` on startup.
     ///
-    /// Called by `AppDelegate+PanelSetup` after verifying that the cached zip
-    /// still exists on disk and the cached version is newer than the installed app.
+    /// Called by `AppUpdater.rehydrateCachedUpdateIfNewer` after verifying the
+    /// cached zip still exists on disk and its version is newer than the running
+    /// app. Clears any stale `updateActionFailed` flag from a prior session so a
+    /// ready-to-install cached update is not masked by the browser fallback.
     public func rehydrateCachedUpdate(zipURL: URL, version: String) {
         updateZipURL = zipURL
         cachedUpdateVersion = version
+        updateActionFailed = false
+    }
+
+    /// Moves to the "downloading" state: clears any cached zip URL / version and
+    /// the failure flag so the host shows a spinner while the new zip downloads.
+    public func setDownloadStarted() {
+        updateZipURL = nil
+        cachedUpdateVersion = nil
+        updateActionFailed = false
+    }
+
+    /// Records a completed, integrity-verified download. The zip is cached at
+    /// `zipURL` for `version`; clears the failure flag so the install
+    /// affordance is shown.
+    public func setDownloadComplete(zipURL: URL, version: String) {
+        updateZipURL = zipURL
+        cachedUpdateVersion = version
+        updateActionFailed = false
+    }
+
+    /// Flags a failed download or install attempt so the browser-download
+    /// fallback is shown.
+    public func setUpdateFailed() {
+        updateActionFailed = true
     }
 
     /// `true` when the latest release exists but its `RunBot.zip` asset is absent.
